# PERF: Performance improvement in to_csv with unused levels in multiindex