--- a/pandas/io/formats/csvs.py
+++ b/pandas/io/formats/csvs.py
@@ -186,6 +186,8 @@ class CSVFormatter:
             data_index = Index(
                 [x.strftime(self.date_format) if notna(x) else "" for x in data_index]
             )
+        elif isinstance(data_index, ABCMultiIndex):
+            data_index = data_index.remove_unused_levels()
         return data_index
 
     @property